move uuid8 functions to main script
