+++ cmem_plugin_uuid/plugin_uuid.py
@@ -1,5 +1,6 @@
 """UUID transform plugin module"""
 
+import os
 import re
 import uuid
 from collections.abc import Sequence
@@ -16,13 +17,47 @@ from cmem_plugin_uuid.utils import (
     get_namespace_uuid,
     namespace_hex,
     node_to_int,
-    parse_uuid8_field,
     repeat_for_inputs,
     uuid3_uuid5_namespace_param,
-    uuid8,
     uuid_convert_param_in,
     uuid_convert_param_out,
 )
+
+
+def uuid8(a: int | None = None, b: int | None = None, c: int | None = None) -> uuid.UUID:
+    """Generate a UUIDv8 from three custom blocks (RFC 9562 §5.8).
+
+    Backport of ``uuid.uuid8`` from the Python 3.14 standard library. Once this
+    project moves to Python 3.14 (when cmem switches), this function is
+    obsolete and callers should use ``uuid.uuid8`` from the stdlib directly.
+
+    * ``a`` is the first 48-bit chunk of the UUID (octets 0-5);
+    * ``b`` is the mid 12-bit chunk (octets 6-7);
+    * ``c`` is the last 62-bit chunk (octets 8-15).
+
+    When a value is not specified, a pseudo-random value is generated.
+
+    The version and variant bits are set manually rather than via
+    ``UUID(version=8)`` because Python's stdlib ``UUID`` constructor rejects
+    versions outside 1-5 prior to Python 3.14.
+    """
+    if a is None:
+        a = int.from_bytes(os.urandom(6))
+    if b is None:
+        b = int.from_bytes(os.urandom(2)) & 0xFFF
+    if c is None:
+        c = int.from_bytes(os.urandom(8)) & 0x3FFFFFFFFFFFFFFF
+    int_uuid_8 = (a & 0xFFFFFFFFFFFF) << 80
+    int_uuid_8 |= (b & 0xFFF) << 64
+    int_uuid_8 |= c & 0x3FFFFFFFFFFFFFFF
+    # Set variant to RFC 4122/9562 ('10' at bits 62-63).
+    int_uuid_8 &= ~(0xC000 << 48)
+    int_uuid_8 |= 0x8000 << 48
+    # Set version 8 (at bits 76-79).
+    int_uuid_8 &= ~(0xF000 << 64)
+    int_uuid_8 |= 8 << 76
+    return uuid.UUID(int=int_uuid_8)
+
 
 _UUID_PATTERN = re.compile(
     r"^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
@@ -357,9 +392,20 @@ class UUID8(TransformPlugin):
     """UUID8 Transform Plugin"""
 
     def __init__(self, a: str = "", b: str = "", c: str = ""):
-        self.a = parse_uuid8_field("a", a, 48) if a else None
-        self.b = parse_uuid8_field("b", b, 12) if b else None
-        self.c = parse_uuid8_field("c", c, 62) if c else None
+        self.a = self._parse_field("a", a, 48) if a else None
+        self.b = self._parse_field("b", b, 12) if b else None
+        self.c = self._parse_field("c", c, 62) if c else None
+
+    @staticmethod
+    def _parse_field(name: str, value: str, bits: int) -> int:
+        """Parse and range-check a UUID8 custom data field."""
+        try:
+            parsed = int(value)
+        except ValueError as exc:
+            raise ValueError(f"{name}: not a valid integer ({value})") from exc
+        if not 0 <= parsed < (1 << bits):
+            raise ValueError(f"{name}: must be a {bits}-bit non-negative integer ({value})")
+        return parsed
 
     def transform(self, inputs: Sequence[Sequence[str]]) -> Sequence[str]:
         """Transform"""
